✨ Integrate PEFT model preparation across trainers and utilities
